Omit `version` from `TagInfo`
The version used to be the `tag_name` minus any prefixes.  But at that
point where we just fetch the tags from a hub, we don't know which
prefixes we could strip -- as that is declared in the registry.

So, we don't make that decision and transformation here but in the
crawler.