Pretend search results are sync
Fixes #42

If the URL does not have a "q", the page shows a statically homepage.
If "q" is set, it enters into search results mode but only after all
the js has loaded.  This results in content flash.

We do two things if "q" is present.

(1) we immediately hide the newest and recent section, but also the
counter.  For the counter, there must be new code that generally undoes
that.  This is done in `setCounter`.  We also fill in the "q" value
into the search input because that looks nice.

(2) we mark the whole document as "bootstrapping", and use that to hide
the "main" and the "footer" section.  We do this until the end of
"index.js" but for not longer than 1 second.

Typically, the results come in fast enough and the page will just
render without any layout jumping or content flashing.  If it takes
more than 1 second, it will basically show pa skeleton page: the input
box will have the correct value, the footer will appear... but nothing
more because of the changes we did in (1).