Handle pre-release versions
The assumption was that we have a 1:1 mapping between release
definitions and resolved releases.  But this is entirely not true in the
presence of pre-releases.  In that case we have a 1:2 mapping, where the
first release, the newer one, is the pre-release and the second is the
final version.

For now, just copy and push a pre-release as a synthetic release into
the `release_definitions`.  Code-wise that is a hack, but the outcome
should be indistinguishable.

Fixes #74