Hide links for removed packages
Either the repos are removed, so we don't need to link to anything
anymore.

Or the author wants to de-list the package for whatever reason.  In the
latter case, usually one would kill the whole package site, but we keep
it for lookup as users might still have that package installed.  Just
keep the info to a minimum to also respect the authors intent.