Unset failing state on successful runs
This was undecided before.  For now: failing state can be reset on
successful runs, the fatal state is final.

We especially need to test if users can undo deleting a repo at all.
If so, it is likely that such repos get a new ID, and thus will throw
`HeartAttack` on the next run.