Allow some basic form of selecting
That is always tricky.  Is the date now a link-alike or still text?
Allow "alt+click" to select just like for links.  Also just check
if something is selected -- which is tricky because
`shadowRoot.getSelection()` is still a draft (but in chromium) according
to an old post on stackoverflow.

After all, double-clicking is still not supported...