Rethink "homepage"-url
Fixes #7

For some GitHub reposiories, we get a "packagecontrol.io" address as
their homepage.  That probably is meant to have back links between
the repo and the pc.io-site.  Drop this if we see it.

Also, refactor the fallback for "homepage".  We actually provide the
repo URL if no "homepage" is explicitly set as before.  Just don't
do this in the "github.py" code but at a later step, when we generate
the channel.