Gracefully handle incomplete packages on the homepage
For the month label of the recent pager on the homepage, we showed
"Jan 1970" for incomplete packages as `timestampValue` defaulted to 0.

This fallback/default is okay to not throw anywhere and for sorting,
but do not show this fake date to the user as the actual updated or
added date.

This is almost straight forward:

a) if the first item of a page is incomplete, all items are because we
are sorted.

b) For the right most item, just search backwards for the first complete
package.  (Usually, the right most package is already complete.)