Fix: in pristine state do not count the totals as the first daily
This is a fix/follow-up to 3a837541 (Fix initial stats delta being 0).

We actually have two states.  When we first build the database (pristine
state) we must not count the totals as the e.g. current daily.  In that
state, we initialize our db with zeros, just what we had in the original
implementation).  In later runs, we might discover new packages, for
these we can count their initial value as the current daily/delta.
(The latter behavior was introduced by 3a837541.)