Fix: On load, prevent any unnecessary `history.push`
Fixes #170

Abort `syncFromUrl` early on page-load if the URL indicates no search.
In that case we would basically run into `revertToNormal` which is the
loading state anyway.  But before that we risk a `history.push` if
`currentPath !== target` (in `goSearch`).

The culprit here is that the currentPath can have additional state on
it coming from the `recent-pager` but the URL we compare to never has.
And that is a bug because `goSearch` now thinks we go to a new page.

Actually we have two ways of preventing a `pushState` here:  (a) by a
switch or (b) by making `currentPath !== target` smarter.

We choose (a) and don't even enter `goSearch`.