Tune platforms lexer to handle "-" special
I.e. we have "-x32" or "-x64" suffixes.  We want the search
"windows-x32" to *not* match also "linux-x32" which would be the case
with the default lexing (splpitting on "-").  But we also want that
"x32" matches both "windows-x32" and "linux-x32" which requires a split
on "-" to work.

Hence we do this on the query parser, and insert "AND" queries if the
search term includes a dash.