Implement time indicators interwoven with the cards
Fixes #135

We split the one big ul and interweave it with `p.timeline-break`
elements.

In general, we iterate over the contents of the whole page

```
    pageItems = this.pagination.calculate()
    buildTimeline(pageItems)
```

and find a bucket size, either "monthly", "quarterly", or finally
"yearly".  We aim to split not into more than 3 groups, unless we're
in "yearly" mode.