Provide ".sublime-package" download links
For the `supportedRepackagerHosts` compute correct `download_url`s
and link to them instead of the raw release urls.

This has the benefit that the user actually downloads a working package,
ready to be dropped into the Installed Packages folder.

The repackager service suggest the correct filename and file extension.
If it suggest ".zip", the package is a no-package package and must be
unzipped into the Packages folder instead.