Implement a way to re-enter auto mode on the labels page
Exactly when the user turns on "omit" *and* the query is empty at the
same time, i.e. the page looks like in pristine state again, also turn
on auto mode again.

Before we turned off auto-mode on first usage of the omit button, and
there was no way to re-enable it.

Ref #242