Fix: ensure onsubmit runs
We catch "enter" in keys.js and move -- as a direct, immediate action --
the focus *away* from the form *to* a card.  As a result, the event
bubbling up will never submit any form as none has the input focus
(anymore).

Since we know that our search results are rendered synchronous, we can
just schedule our focus move / side-effect for the next task.