Disable omit toggle while searching
Fixes #242
Closes #243

This effectively enforces the `autoMode` and disallows toggling the omit
state while in a search.

Since we don't do anything special when showing search results, the
explaining sentence "Omitting ..." is slowly faded away.

The implementation reduces the state management to two attributes on
the element, namely "disabled" and "aria-pressed".  The elements state
is thus the source of truth.